major refac: moved major parent components to pages section
