+++ src/components/index.js
@@ -1,4 +1,3 @@
 export { Header } from "./Header/Header";
 export { SideBar } from "./SideBar/SideBar";
-export { DisplayGrid } from "./DisplayGrid/DisplayGrid";
-export { VideoPage } from "./VideoPage/VideoPage";
+export { Thumbnail } from "./Thumbnail/Thumbnail";
