+++ src/pages/index.js
@@ -1,1 +1,3 @@
 export { Home } from "./Home/Home";
+export { DisplayGrid } from "./DisplayGrid/DisplayGrid";
+export { VideoPage } from "./VideoPage/VideoPage";
